Refactor to separate provider and connector classes
